requirements.txt
# Core Dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
streamlit>=1.12.0
pydantic>=1.10.0
python-multipart>=0.0.5

# ML and Data Processing
numpy>=1.23.0,<2.0.0
scikit-learn>=1.3.0
pandas>=1.5.0
joblib>=1.1.0

# AWS and Bedrock
boto3>=1.26.0
botocore>=1.29.0

# Testing
hypothesis>=6.50.0
pytest>=7.0.0
pytest-asyncio>=0.20.0

# RAG System Dependencies
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
pypdf>=3.17.4
tiktoken>=0.5.2

# Utilities
python-dotenv>=0.19.0
requests>=2.27.0
